fix: fixed circular deps
