+++ test/e2e/tests/settings/state-logs.json
@@ -641,6 +641,7 @@
     },
     "completedMetaMetricsOnboarding": "boolean",
     "completedOnboarding": "boolean",
+    "hasSeenOnboardingCompletionPage": "boolean",
     "configs": {
       "networks": "object"
     },
