+++ ui/selectors/first-time-flow.js
@@ -1,11 +1,6 @@
 import { FirstTimeFlowType } from '../../shared/constants/onboarding';
 import { getIsSeedlessOnboardingFeatureEnabled } from '../../shared/lib/environment';
 import {
-  getIsInitialized,
-  getIsWalletResetInProgress,
-} from '../ducks/metamask/metamask';
-import { getIsUnlocked } from '../ducks/metamask/base-selectors';
-import {
   DEFAULT_ROUTE,
   ONBOARDING_COMPLETION_ROUTE,
   ONBOARDING_CREATE_PASSWORD_ROUTE,
@@ -53,26 +48,6 @@ export function getFirstTimeFlowTypeRouteAfterUnlock(state) {
     return ONBOARDING_DOWNLOAD_APP_ROUTE;
   }
   return DEFAULT_ROUTE;
-}
-
-/**
- * True when a locked user should unlock before resuming the onboarding
- * completion page (return visit without tapping Done).
- *
- * @param {object} state - MetaMask state tree
- * @returns {boolean} Whether the user must unlock before onboarding completion
- */
-export function getShouldUnlockBeforeOnboardingCompletion(state) {
-  const { hasSeenOnboardingCompletionPage, completedOnboarding } =
-    state.metamask;
-
-  return (
-    hasSeenOnboardingCompletionPage &&
-    !completedOnboarding &&
-    !getIsUnlocked(state) &&
-    getIsInitialized(state) &&
-    !getIsWalletResetInProgress(state)
-  );
 }
 
 /**
